images:fetch_representatives now skips download if the file exists. Use FORCE=1 to override.

+++ lib/tasks/images.rake
@@ -10,6 +10,11 @@ namespace :images do
       url = URI.parse("http://stortinget.no/Personimages/PersonImages_ExtraLarge/#{URI.escape rep.external_id}_ekstrastort.jpg")
 
       filename = rep_image_path.join("#{rep.slug}.jpg")
+      
+      if ENV['FORCE'].nil? && filename.exist?
+        puts "skipping download for existing #{filename}, use FORCE=true to override"
+        next
+      end
 
       File.open(filename.to_s, "wb") do |destination|
         resp = Net::HTTP.get_response(url) do |response|
